+++ src/main/java/com/backend/rentacarbd/domain/CarDto.java
@@ -17,10 +17,10 @@ public class CarDto {
     private String engineType;
     private Integer engineCapacity;
     private Integer productionYear;
-    private Double costPerDay;
+    private Integer costPerDay;
     private boolean available;
 
-    public CarDto(String brand, String model, String colour, String engineType, Integer engineCapacity, Integer productionYear, Double costPerDay, boolean available) {
+    public CarDto(String brand, String model, String colour, String engineType, Integer engineCapacity, Integer productionYear, Integer costPerDay, boolean available) {
         this.brand = brand;
         this.model = model;
         this.colour = colour;
